chore: refactoring for notification status and pagination utility
