+++ packages/pn-commons/src/index.tsx
@@ -7,6 +7,14 @@ import CustomPagination from './components/Pagination/CustomPagination';
 import CustomTooltip from './components/CustomTooltip';
 import AppMessage from './components/AppMessage';
 
+export { LoadingOverlay };
+export { Header };
+export { Layout };
+export { Footer };
+export { CustomPagination };
+export { CustomTooltip };
+export { AppMessage };
+
 // pages
 import NotFound from './navigation/NotFound';
 import AccessDenied from './navigation/AccessDenied';
@@ -17,23 +25,22 @@ export { AccessDenied };
 // types
 import { AppError } from './types/AppError';
 import { PaginationData } from './components/Pagination/types';
-import { createAppError } from './services/error.service';
-import { formatDate } from './services/date.service';
-
-// components
-export { LoadingOverlay };
-export { Header };
-export { Layout };
-export { Footer };
-export { CustomPagination };
-export { CustomTooltip };
-export { AppMessage };
-
-// types
+import { NotificationStatus } from './types/NotificationStatus';
 export type { AppError };
 export type { PaginationData };
+export { NotificationStatus };
 
 // functions
+import { createAppError } from './services/error.service';
+import { formatDate } from './services/date.service';
+import { calcPages } from './utils/pagination.utility';
+import {
+  getNotificationStatusLabelAndColor,
+  NotificationAllowedStatus,
+} from './utils/status.utility';
 
+export { NotificationAllowedStatus };
+export { getNotificationStatusLabelAndColor };
 export { createAppError };
 export { formatDate };
+export { calcPages };
