+++ packages/pn-commons/src/types/NotificationStatus.ts
@@ -1,0 +1,12 @@
+export enum NotificationStatus {
+  IN_VALIDATION = 'IN_VALIDATION',
+  ACCEPTED = 'ACCEPTED',
+  REFUSED = 'REFUSED',
+  DELIVERING = 'DELIVERING',
+  DELIVERED = 'DELIVERED',
+  VIEWED = 'VIEWED',
+  EFFECTIVE_DATE = 'EFFECTIVE_DATE',
+  PAID = 'PAID',
+  UNREACHABLE = 'UNREACHABLE',
+  CANCELED = 'CANCELED',
+}
