+++ packages/pn-commons/src/utils/__test__/status.utility.test.ts
@@ -1,5 +1,5 @@
-import { NotificationStatus } from '../../redux/dashboard/types';
 
+import { NotificationStatus } from '../../types/NotificationStatus';
 import { getNotificationStatusLabelAndColor } from '../status.utility';
 
 function testStatusLabelAndColorFn(
