+++ packages/pn-pa-webapp/src/api/notifications/Notifications.api.ts
@@ -1,7 +1,7 @@
+import { formatDate } from '@pagopa-pn/pn-commons';
 import { apiClient } from '../axios';
 import { GetNotificationsParams, GetNotificationsResponse } from '../../redux/dashboard/types';
 import { NotificationDetail, Legalfacts } from '../../redux/notification/types';
-import { formatDate } from '@pagopa-pn/pn-commons';
 
 export const NotificationsApi = {
   /**
