+++ packages/pn-pa-webapp/src/redux/dashboard/__test__/reducers.test.ts
@@ -1,4 +1,4 @@
-import { formatDate } from '@pagopa-pn/pn-commons';
+import { formatDate, NotificationStatus } from '@pagopa-pn/pn-commons';
 import MockAdapter from 'axios-mock-adapter';
 
 import { apiClient } from '../../../api/axios';
@@ -7,7 +7,7 @@ import { exchangeToken, logout } from '../../auth/actions';
 import { loginInit } from '../../auth/__test__/reducers.test';
 import { store } from '../../store';
 import { getSentNotifications, setNotificationFilters, setPagination, setSorting } from '../actions';
-import { GetNotificationsResponse, NotificationStatus } from '../types';
+import { GetNotificationsResponse } from '../types';
 
 const notificationsFromBe: GetNotificationsResponse = {
   result: [
