+++ packages/pn-pa-webapp/src/redux/dashboard/types.ts
@@ -1,3 +1,5 @@
+import { NotificationStatus } from "@pagopa-pn/pn-commons";
+
 export interface Notification {
     iun: string;
     paNotificationId: string;
@@ -7,20 +9,6 @@ export interface Notification {
     notificationStatus: NotificationStatus;
     recipientId: string;
 }
-
-export enum NotificationStatus {
-    IN_VALIDATION = 'IN_VALIDATION',
-    ACCEPTED = 'ACCEPTED',
-    REFUSED = 'REFUSED',
-    DELIVERING = 'DELIVERING',
-    DELIVERED = 'DELIVERED',
-    VIEWED = 'VIEWED',
-    EFFECTIVE_DATE = 'EFFECTIVE_DATE',
-    PAID = 'PAID',
-    UNREACHABLE = 'UNREACHABLE',
-    CANCELED = 'CANCELED'
-}
-
 export interface GetNotificationsResponse {
     result: Array<Notification>;
     moreResult: boolean;
