+++ packages/pn-personafisica-webapp/src/api/notifications/Notifications.api.ts
@@ -1,9 +1,10 @@
+import { formatDate } from '@pagopa-pn/pn-commons';
+
 import { apiClient } from '../axios';
 import {
   GetNotificationsParams,
   GetNotificationsResponse,
 } from '../../redux/dashboard/types';
-import { formatDate } from '@pagopa-pn/pn-commons';
 
 export const NotificationsApi = {
   /**
